Got uros publishing without crashing when agent goes up and down, but this is without uros options that have domain ID

+++ hardware/firmware/clearcore_drive_motors/lib/uros/uros_config.h
@@ -7,7 +7,7 @@
 #define FIRMWARE_VERSION "0.1.0"
 #define HARDWARE_ID "1ceec360-32b9-431c-9166-1252809e85b2"
 
-constexpr int8_t kDomainId = 10;
+constexpr size_t kDomainId = 10;
 constexpr char kNodeName[] = "uros_clearcore_client";
 constexpr char kNamespace[] = "";
